outsources websocketListener into a separate class, adds request for complete blockdata.
